+++ docs/source/index.rst
@@ -1,6 +1,15 @@
 **dgbowl-schemas**: schemas and validators for the dgbowl suite of tools
 ========================================================================
 
+This repository contains the versioned archive of the schemas for the following tools:
+
+- The :class:`~dgbowl_schemas.dgpost.Recipe` schema for :mod:`dgpost`, with its latest
+  implementation in :mod:`~dgbowl_schemas.dgpost.recipe`;
+- The :class:`~dgbowl_schemas.yadg.DataSchema` schema for :mod:`yadg`, with its latest
+  implementation in :mod:`~dgbowl_schemas.yadg.dataschema`;
+- The :class:`~dgbowl_schemas.tomato.Recipe` schema for :mod:`tomato`, with its latest
+  implementation in :mod:`~dgbowl_schemas.tomato.recipe`.
+
 .. toctree::
    :maxdepth: 1
    :caption: dgbowl-schemas autodocs
