+++ src/dgbowl_schemas/dgpost/__init__.py
@@ -1,12 +1,11 @@
 import logging
+from . import recipe
 from pydantic import ValidationError
-from . import recipe_2_1
 from .recipe_2_1 import Recipe as Recipe_2_1
 from .recipe_1_0 import Recipe as Recipe_1_0
 
 logger = logging.getLogger(__name__)
 
-recipe = recipe_2_1
 Recipe = Recipe_2_1
 
 models = {
@@ -27,3 +26,10 @@ def to_recipe(**kwargs):
             if firste is None:
                 firste = e
     raise ValueError(firste)
+
+
+__all__ = [
+    "recipe",
+    "Recipe",
+    "to_recipe",
+]
