+++ src/dgbowl_schemas/yadg/__init__.py
@@ -1,6 +1,6 @@
 from pydantic import ValidationError
 import logging
-from . import dataschema_5_0
+from . import dataschema
 from .dataschema_5_0 import (
     DataSchema as DataSchema_5_0,
     Metadata as Metadata_5_0,
@@ -14,11 +14,7 @@ from .dataschema_4_0 import DataSchema as DataSchema_4_0, Metadata as Metadata_4
 
 logger = logging.getLogger(__name__)
 
-dataschema = dataschema_5_0
 DataSchema = DataSchema_5_0
-Metadata = Metadata_5_0
-FileType = FileType_5_0
-ExtractorFactory = ExtractorFactory_5_0
 
 models = {
     "5.0": (DataSchema_5_0, Metadata_5_0),
@@ -50,3 +46,10 @@ def to_dataschema(**kwargs):
     # attempt to parse using identified Model
     schema = Model(**kwargs)
     return schema
+
+
+__all__ = [
+    "dataschema",
+    "DataSchema",
+    "to_dataschema",
+]
