+++ src/dgbowl_schemas/yadg/dataschema/__init__.py
@@ -1,0 +1,9 @@
+from ..dataschema_5_0 import (
+    DataSchema,
+    Metadata,
+    StepDefaults,
+    FileType,
+    ExtractorFactory,
+)
+
+__all__ = ["DataSchema", "Metadata", "StepDefaults", "FileType", "ExtractorFactory"]
